# Optimized sequence encoding for scalars